requirements.txt
numpy
pandas
Pillow
tifffile
torch>=1.13
monai
scipy
scikit-image
connected-components-3d
tqdm
pyyaml
nbconvert
nbformat
pytest
